# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Home AI OS",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/sharing/rules/{category}/{member_id}": {
      "delete": {
        "tags": [
          "持续共享"
        ],
        "summary": "Change",
        "operationId": "change_api_v1_sharing_rules__category___member_id__delete",
        "parameters": [
          {
            "name": "category",
            "in": "path",
            "required": true,
            "schema": {
              "enum": [
                "health",
                "location"
              ],
              "type": "string",
              "title": "Category"
            }
          },
          {
            "name": "member_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Member Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "持续共享"
        ],
        "summary": "Change",
        "operationId": "change_api_v1_sharing_rules__category___member_id__put",
        "parameters": [
          {
            "name": "category",
            "in": "path",
            "required": true,
            "schema": {
              "enum": [
                "health",
                "location"
              ],
              "type": "string",
              "title": "Category"
            }
          },
          {
            "name": "member_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Member Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/data/{record_id}/visibility": {
      "put": {
        "tags": [
          "个人与家庭范围"
        ],
        "summary": "Visibility",
        "operationId": "visibility_api_v1_data__record_id__visibility_put",
        "parameters": [
          {
            "name": "record_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Record Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VisibilityInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/sharing/categories/{category}": {
      "put": {
        "tags": [
          "个人与家庭范围"
        ],
        "summary": "Category",
        "operationId": "category_api_v1_sharing_categories__category__put",
        "parameters": [
          {
            "name": "category",
            "in": "path",
            "required": true,
            "schema": {
              "enum": [
                "health",
                "location"
              ],
              "type": "string",
              "title": "Category"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VisibilityInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/tasks": {
      "post": {
        "summary": "Create Task",
        "operationId": "create_task_api_v1_tasks_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tasks/{task_id}": {
      "get": {
        "summary": "Get Task",
        "operationId": "get_task_api_v1_tasks__task_id__get",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Predicate": {
        "properties": {
          "step": {
            "type": "integer",
            "maximum": 15.0,
            "minimum": 0.0,
            "title": "Step"
          },
          "path": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "maxItems": 20,
            "title": "Path"
          },
          "operator": {
            "type": "string",
            "enum": [
              "exists",
              "not_exists",
              "not_empty",
              "equals",
              "not_equals",
              "gt",
              "gte",
              "lt",
              "lte"
            ],
            "title": "Operator"
          },
          "value": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "integer"
              },
              {
                "type": "number"
              },
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Value"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "step",
          "operator"
        ],
        "title": "Predicate"
      },
...
      "Step": {
        "properties": {
          "capability": {
            "type": "string",
            "title": "Capability"
          },
          "arguments": {
            "additionalProperties": true,
            "type": "object",
            "title": "Arguments"
          },
          "when": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/StepCondition"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "capability"
        ],
        "title": "Step"
      },
      "StepCondition": {
        "properties": {
          "mode": {
            "type": "string",
            "enum": [
              "all",
              "any"
            ],
            "title": "Mode",
            "default": "all"
          },
          "predicates": {
            "items": {
              "$ref": "#/components/schemas/Predicate"
            },
            "type": "array",
            "maxItems": 16,
            "minItems": 1,
            "title": "Predicates"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "predicates"
        ],
        "title": "StepCondition"
      },
...
      "TaskRequest": {
        "properties": {
          "message": {
            "type": "string",
            "maxLength": 20000,
            "title": "Message",
            "default": ""
          },
          "idempotency_key": {
            "type": "string",
            "maxLength": 200,
            "minLength": 8,
            "title": "Idempotency Key"
          },
          "mode": {
            "type": "string",
            "enum": [
              "local",
              "cloud"
            ],
            "title": "Mode",
            "default": "local"
          },
          "record_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 20,
            "title": "Record Ids"
          },
          "capability": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Capability"
          },
          "arguments": {
            "additionalProperties": true,
            "type": "object",
            "title": "Arguments"
          },
          "max_steps": {
            "type": "integer",
            "maximum": 16.0,
            "minimum": 1.0,
            "title": "Max Steps",
            "default": 8
          },
          "max_output_tokens": {
            "type": "integer",
            "maximum": 4096.0,
            "minimum": 1.0,
            "title": "Max Output Tokens",
            "default": 1024
          },
          "steps": {
            "items": {
              "$ref": "#/components/schemas/Step"
            },
            "type": "array",
            "maxItems": 16,
            "title": "Steps"
          },
          "timeout_seconds": {
            "type": "integer",
            "maximum": 3600.0,
            "minimum": 10.0,
            "title": "Timeout Seconds",
            "default": 600
          },
          "step_timeout_seconds": {
            "type": "integer",
            "maximum": 300.0,
            "minimum": 1.0,
            "title": "Step Timeout Seconds",
            "default": 120
          },
          "max_model_tokens": {
            "type": "integer",
            "maximum": 262144.0,
            "minimum": 256.0,
            "title": "Max Model Tokens",
            "default": 32768
          },
          "max_read_retries": {
            "type": "integer",
            "maximum": 3.0,
            "minimum": 0.0,
            "title": "Max Read Retries",
            "default": 1
          },
          "timezone": {
            "type": "string",
            "maxLength": 100,
            "title": "Timezone",
            "default": "Asia/Shanghai"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "idempotency_key"
        ],
        "title": "TaskRequest"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "VisibilityInput": {
        "properties": {
          "visibility": {
            "type": "string",
            "enum": [
              "personal",
              "family"
            ],
            "title": "Visibility"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "visibility"
        ],
        "title": "VisibilityInput"
      },
...
    }
  }
}